// src/common/enums/code.enum.ts
export enum Code {
    UNAUTHORIZED = 'UNAUTHORIZED',
    BAD_REQUEST = 'BAD_REQUEST',
    NOT_FOUND = 'NOT_FOUND',
    CONFLICT = 'CONFLICT',
    FORBIDDEN = 'FORBIDDEN',
    TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
    INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
    SERVICE_UNAVAILABLE = 'SERVICE_UNAVAILABLE',
    // gql specific
    GRAPHQL_VALIDATION_FAILED = 'GRAPHQL_VALIDATION_FAILED',
    BAD_USER_INPUT = 'BAD_USER_INPUT',
}